* src/ImageSelect.py Fixed typo in on_drag_begin
* src/RelLib.py
Check for the existance of the witness member of an Event. ZODB
stores an object, so elements read in from old databases may not
have the element, which can cause a traceback.


* src/AutoComp.py
Don't build popdown lists if the number of elements exceeds 250.
Otherwise, the menus will become unusable, and just consume
tons of memory.

* src/VersionControl.py
Sort revision numbers based on numerical value instead of ASCII value.
* src/VersionControl.py
Check return values of all subprocesses


git-svn-id: svn+ssh://svn.code.sf.net/p/gramps/code/trunk/gramps2@1663 4ae1f11a-8b86-4847-b8af-ab372f36d1fd