4198: Person view does not remove a row correctly when two people are merged.
This is a major patch by Michael Nauta. It means all writes happen immediately to bsddb, and the bsddb 
rollback is used on a crash. Transaction is no longer used to store changes and do them on commit.
Undo database is set on end. 
At the same time with statement is used throughout for transactions
At the same time, the original bug in merge code should be fixed
Still some issues, that will be ironed out


git-svn-id: svn+ssh://svn.code.sf.net/p/gramps/code/trunk@16523 4ae1f11a-8b86-4847-b8af-ab372f36d1fd