[r21613]Determine Paths at Runtime
Uses GRAMPS_RESOURCES to override the root location of the Gramps data
files (which would normally be $(prefix)/share). In setup.py install,
writes $(prefix)/share to gen/utils/resource-path, installs that file,
then deletes it again from the source tree. The presence or absence
determines whether Gramps is running from the source directory or from
an installation.

Const.py is now a static file; const.py.in is no longer used. Note that
because importing const into setup tried to initialize things that we
don't want initialized, VERSION is removed, and imported from
gramps.version -- a file which will be added in the next change.
Consequently, this commit will not run.

git-svn-id: svn+ssh://svn.code.sf.net/p/gramps/code/branches/maintenance/gramps40@21618 4ae1f11a-8b86-4847-b8af-ab372f36d1fd